Requires user to set the repo endpoint url
The repo endpoint is used to construct the baseurl in the yum
repo definitions. Previously it was assumed the endpoint would
always be S3. However, it is useful to front S3 with Cloudfront
or another CDN, at which point the endpoint can be most any URL.

To restore the prior behavior, set:

```
repo_endpoint = <s3_endpoint>/<bucket_name>
```