Supports creating multiple salt repos at different repo_prefixes
This better supports the workflow when saltstack moves a version
from the main repo to the archive repo.

Previously, a user would need to invoke this module once for the
versions in the main repo, and once for the versions in the archive
repo. This could lead to race conditions, particularly on the yum
definition files, when using the same yum_prefix for main and archive
versions, i.e. as one call to the module is destroying the yum
definition for a given version, the other call is creating the same
yum definition.

This patch moves all the yum definitions into the same aws_s3_bucket_object
resource, so terraform can properly managed the object and its content,
without a race condition.