Adds feature to specify the source bucket name for the rclone command
The cloudfront cache in front of the salt archive repo is currently
experiencing some issues that prevent the bucket listing from working
correctly. This patchset provides an option for setting the "bucket name"
in the rclone command, so the user can hit the underlying s3 bucket
directly, instead of going through cloudfront.

It is unfortunate that this is backwards-incompatible, in that due
to the way terraform objects work, the user *must* add the new attribute
to their input objects.